requirements.txt
arrow==1.3.0
asgiref==3.7.2
binaryornot==0.4.4
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cookiecutter==2.4.0
crispy-tailwind==0.5.0
Django==4.2.7
django-crispy-forms==2.1
django-tailwind==3.6.0
gunicorn==21.2.0
idna==3.4
Jinja2==3.1.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
packaging==23.2
Pillow==10.1.0
psycopg==3.1.12
psycopg2-binary==2.9.9
Pygments==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
PyYAML==6.0.1
requests==2.31.0
rich==13.6.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
urllib3==2.0.7
whitenoise==6.6.0
